Leave request

+++ LeaveRequest/Services/Implements/LeaveRequestService.cs
@@ -7,6 +7,8 @@ using LeaveRequest.DTO.Leave;
 using LeaveRequest.Models;
 using Microsoft.EntityFrameworkCore;
 using Shared;
+using Shared.Cache;
+using Shared.Data;
 using Shared.Exceptions;
 
 namespace LeaveRequest.Services.Implements
@@ -15,18 +17,39 @@ namespace LeaveRequest.Services.Implements
     {
         private readonly LeaveRequestServiceContext _context;
         private readonly IMapper _mapper;
+        private readonly IScopedCache _cache;
 
-        public LeaveRequestService(LeaveRequestServiceContext context, IMapper mapper)
+        public LeaveRequestService(LeaveRequestServiceContext context, IMapper mapper, IScopedCache cache)
         {
             _context = context;
             _mapper = mapper;
+            _cache = cache;
         }
 
         //get leave request
-        public async Task<List<LeaveRequestResponse>> GetLeaveRequestAsync()
+        public async Task<List<LeaveRequestResponse>> GetLeaveRequestAsync(LeaveRequestListRequest b)
         {
-            var leave = await _context.LeaveRequests.Include(b => b.LeaveTypes).ToListAsync();
-
+            var leave = new List<Request>();
+     
+            if (b.LeaveDates != null)
+            {
+                leave = await _context.LeaveRequests.Include(b => b.LeaveTypes).Where(x => x.LeaveDates == b.LeaveDates).ToListAsync();
+            }
+            
+            if (b.LeaveTime != null)
+            {
+                leave = await _context.LeaveRequests.Include(b => b.LeaveTypes).Where(x => x.LeaveTime == b.LeaveTime).ToListAsync();
+            }
+            
+            if (b.LeaveTypeId != 0)
+            {
+                leave = await _context.LeaveRequests.Include(x => x.LeaveTypes).Where(x => x.LeaveTypeId == b.LeaveTypeId).ToListAsync();
+            }
+            else if((b.LeaveTime == null) && (b.LeaveTypeId == 0) && (b.LeaveDates == null))
+            {
+                leave = await _context.LeaveRequests.Include(b => b.LeaveTypes).ToListAsync();
+            }
+            
             return _mapper.Map<List<LeaveRequestResponse>>(leave);
         }
 
@@ -34,6 +57,11 @@ namespace LeaveRequest.Services.Implements
         public async Task DeleteLeaveRequestAsync(int id)
         {
             var leave = await _context.LeaveRequests.FindAsync(id);
+            
+            if (leave.MemberEmail != _cache.Email && _cache.Role != Constrain.AdminRole)
+            {
+                throw new ServiceException(401, "You don't have permission to delete this Booking");
+            }
 
             _context.LeaveRequests.Remove(leave);
             await _context.SaveChangesAsync();
@@ -42,12 +70,29 @@ namespace LeaveRequest.Services.Implements
         //create leave request
         public async Task<LeaveRequestCreateResponse> CreateLeaveRequestAsync(LeaveRequestCreateRequest request)
         {
-            var leave = _mapper.Map<Request>(request);
+            foreach(var date in request.LeaveDates)
+            {
+                var entity = new Request
+                {
+                    LeaveDates = date,
+                    LeaveTime = request.LeaveTime,
+                    LeaveTypeId = request.LeaveTypeId,
+                    Name = request.Name,
+                    Reason = request.Reason,
+                    MemberEmail = _cache.Email,
+                    MemberName = _cache.Username
+                };
+                _context.LeaveRequests.Add(entity);
+            }
 
-            _context.LeaveRequests.Add(leave);
             await _context.SaveChangesAsync();
 
-            return _mapper.Map<LeaveRequestCreateResponse>(leave);
+            return _mapper.Map<LeaveRequestCreateResponse>(new LeaveRequestCreateResponse());
+        }
+
+        public Task<List<LeaveRequestResponse>> GetLeaveRequestAsync(LeaveRequestCreateRequest request)
+        {
+            throw new NotImplementedException();
         }
     }
 }
